chore(tests): move test files to __tests__ folder
