+++ package/src/functions/__tests__/forward-params.test.ts
@@ -1,7 +1,7 @@
 // tslint:disable:max-line-length
 import { Routes } from '@angular/router';
-import { forwardParams } from './forward-params';
-import { createRoot } from '../creators/root.creator';
+import { forwardParams } from '../forward-params';
+import { createRoot } from '../../creators/root.creator';
 
 describe('forwardParams', () => {
   it('should return the same state', () => {
